finish pretty-printing refactor
