+++ app/game/Swarm/App.hs
@@ -34,7 +34,7 @@ import Graphics.Vty qualified as V
 import Graphics.Vty.CrossPlatform qualified as V
 import Swarm.Game.Failure (SystemFailure)
 import Swarm.Game.State.Runtime
-import Swarm.Language.Pretty (prettyText)
+import Swarm.Pretty (prettyText)
 import Swarm.Log (LogSource (SystemLog), Severity (..))
 import Swarm.TUI.Controller
 import Swarm.TUI.Model
