+++ src/swarm-lang/Swarm/Language/Syntax.hs
@@ -102,5 +102,6 @@ import Swarm.Language.Syntax.Comments
 import Swarm.Language.Syntax.Constants
 import Swarm.Language.Syntax.Loc
 import Swarm.Language.Syntax.Pattern
+import Swarm.Language.Syntax.Pretty ()
 import Swarm.Language.Syntax.Util
 import Swarm.Language.Types
