+++ src/swarm-scenario/Swarm/Game/World/Load.hs
@@ -21,7 +21,7 @@ import Swarm.Game.Land
 import Swarm.Game.ResourceLoading (getDataDirSafe)
 import Swarm.Game.World.Parse (parseWExp, runParser)
 import Swarm.Game.World.Typecheck
-import Swarm.Language.Pretty (prettyText)
+import Swarm.Pretty (prettyText)
 import Swarm.Util (acquireAllWithExt)
 import Swarm.Util.Effect (throwToWarning, withThrow)
 import System.FilePath (dropExtension, joinPath, splitPath)
