+++ src/swarm-tui/Swarm/TUI/Controller/UpdateUI.hs
@@ -33,7 +33,7 @@ import Swarm.Game.Robot.Concrete
 import Swarm.Game.State
 import Swarm.Game.State.Landscape
 import Swarm.Game.State.Substate
-import Swarm.Language.Pretty
+import Swarm.Pretty
 import Swarm.Language.Typed (Typed (..))
 import Swarm.Language.Types
 import Swarm.Language.Value (Value (VExc, VUnit), envTydefs, prettyValue)
