+++ src/swarm-tui/Swarm/TUI/Launch/Model.hs
@@ -18,7 +18,7 @@ import Data.Text (Text)
 import Swarm.Game.Failure (SystemFailure)
 import Swarm.Game.Scenario.Status (ParameterizableLaunchParams (LaunchParams), ScenarioInfoPair, SerializableLaunchParams)
 import Swarm.Game.State (LaunchParams, ValidatedLaunchParams, getRunCodePath, parseCodeFile)
-import Swarm.Language.Pretty (prettyText)
+import Swarm.Pretty (prettyText)
 import Swarm.TUI.Model.Name
 import Swarm.Util.Effect (withThrow)
