+++ src/swarm-tui/Swarm/TUI/View/Util.hs
@@ -24,7 +24,7 @@ import Swarm.Game.State
 import Swarm.Game.State.Landscape
 import Swarm.Game.State.Substate
 import Swarm.Game.Terrain
-import Swarm.Language.Pretty (prettyTextLine)
+import Swarm.Pretty (prettyTextLine)
 import Swarm.Language.Syntax (Syntax)
 import Swarm.Language.Text.Markdown qualified as Markdown
 import Swarm.Language.Types (Polytype)
