+++ src/swarm-web/Swarm/Web.hs
@@ -76,7 +76,7 @@ import Swarm.Game.State.Robot
 import Swarm.Game.State.Substate
 import Swarm.Game.Step.Path.Type
 import Swarm.Language.Pipeline (processTermEither)
-import Swarm.Language.Pretty (prettyTextLine)
+import Swarm.Pretty (prettyTextLine)
 import Swarm.TUI.Model hiding (SwarmKeyDispatchers (..))
 import Swarm.TUI.Model.Dialog.Goal
 import Swarm.TUI.Model.Repl (REPLHistItem, replHistory, replSeq)
